+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 * Added search functionality for sheets in admin.
 * Add one-handed use penalty for firearms, big pistols and such are hard to aim with only one hand.
+* Add overall damage reduction to armor.
 
 ## [0.23.0]